Import of SquidLib RNG code.
Some small changes: RNG's shuffle() has changed from a GwtIncompatible
out-of-place shuffle (no GWT means it is unused within SquidLib) to an
in-place shuffle that is compatible. Seeding an RNG variant with a
String gets a 64-bit hash instead of a 32-bit hash, since most seeds are
64-bit here anyway. RNG methods that took a ShortSet are removed (no
ShortSet class here).