0.1; brings in updated code from SquidLib
CrossHash changes its algorithm here, so string hashing will be about 7x
faster and less predictable relative to input than with FNV-1a, but
string seeds used with the (barely available) previous version won't
have the same results. IsaacRNG is a new implementation of an old but
good algorithm that is somewhere between crypto-ish PRNGs and non-crypto
PRNGs. ChaosRNG is removed because SecureRandom is unlikely to be
present on many non-desktop platforms and it didn't provide much on top
of a SecureRandom seed for a LongPeriodRNG.