Version 0.2; adds ThunderRNG, improves CrossHash
ThunderRNG is new, from the latest SquidLib code, and it is yet to be
seen if its statistical qualities are good enough for uses that require
ideal distribution, but its speed is top-notch, faster than all the
other RandomnessSource implementations.

ALSO: Update a test, CrossHash, SobolQRNG

SobolQRNG got some changes to a few methods to avoid them generating
all-0 in cases they shouldn't, and added nextDouble() so it can be
tested like actual PRNGs. CrossHash updated to have the same
improvements just introduced to SquidLib's Lightning hash algorithm, and
we might as well update it in this early version before anything else
depends on CrossHash having fixed outputs for the same inputs. The
improvements are mainly a fix in a bug-like quirk where the highest bits
of a 64-bit hash from CrossHash weren't filled as unpredictably as they
should. Specifically, on 2-element arrays, a carefully chosen second
element could heavily bias the hash to have mostly or all 0 for the
higher bits (lower bits were unaffected), regardless of the first
element; this has been fixed by switching the finalization step from
XOR-ing with two shifted versions of an otherwise-unchanged variable to
XOR-ing once with a bitwise-rotated version of a heavily-modified
variable. Speed is almost the same, just a tiny amount slower because
the finalization step is only run once regardless of the size of the
data being hashed.