Package changes, CrossHash.Storm quality up
The package needed to change (from sarong.rng to sarong, specifically,
with sarong.util staying where it is) for GWT to be more likely to work
out of the box with Sarong. The .gwt.xml file was in the wrong place to
be used from actual GWT projects, it seems. CrossHash.Storm now factors
in the salt more heavily during hashing elements of the input, as
opposed to just at the start and end. Compatibility has a little more
doumentation added. Both types of CrossHash now allow hashing general
Object arrays, using each Object's hashCode() as an element of input.
There's also a benchmarks module that shouldn't be considered part of
the main code, and doesn't add to dependencies or jar size.