Update to 0.4 with more stuff
CrossHash.Falcon is faster but has some limitations. PintRNG is slower
than other RNGs on most platforms, but should be faster on GWT.
VanDerCorputQRNG is essentially an alternative to SobolQRNG for
quasi-random number sequences, particularly in 2D or 3D.
Compatibility.first() is irrelevant here and was removed. Some package
changes were introduced in the last commit, and are packagees are
different from 0.3 to better support GWT.