Clean everything up; 0.4.1 release
There was some messy code all over Sarong, and this release attempts to
clean up as much as I could find. In particular, equals() and hashCode()
methods are present on all RandomnessSource implementations now, which
should be useful, and determine() static methods are more common to
allow usage of a RandomnessSource's algorithm without actually affecting
a state (these usually take a long as input and produce another
random-seeming long as output). All of the nextInt() methods in
RandomnessSource implementations now use the same technique as in RNG
(Lemire's way to avoid modulus), and they may be slightly faster without
the RNG wrapper. JavaRNG is new, and is a simple way to compare or test
against java.util.Random using RNG; it is Ben McLean's work and my only
improvements were adding simple equals() and hashCode() implementations.
PintRNG and PermutedRNG should be faster and produce different random
numbers now.