Figured out that Slap was collapsing its cycle
After collapsing, which happened in the previous SlapRNG at some point
in the first 2 to the 34 generations, the rest of the period would
reduce to some smaller value, usually 2 to the 32, which is not what we
want. Now we have a probably-longer-period SlapRNG that's a fair amount
slower, though quality seems good.