Change Herd/Horde to avoid chance of collapse
There was a remote possibility that the earlier Herd and Horde
algorithms could wind up in a state where no changes would happen after
the state reached a full-standstill position. This seems to be remedied
in the latest code by adding a constant that will always change at least
one bit in state, breaking free of standstill. Speed is a little lower,
but a full period without collapsing into a short repetitive pattern is
much more important!