BirdRNG, for all your Bird Law (and other) needs
BirdRNG passes PractRand with 64 MB of random ints, for everything but
32-bit folding. It passes with no anomalies for normal folding (stdin),
8-bit (stdin8) and 16-bit (stdin16), and with no failures but several
anomalies with 64-bit folding (stdin64). It has two failures on 32-bit
folding; this may be possible to remedy later. It is currently the
RandomnessSource with the longest period of all 32-bit generators, and
though it could be as low as 2 to the 512 or even lower, it would still
be higher than HerdRNG. The quality is big, though. It isn't especially
slow at int generation, either.