Increase state size of BirdRNG; it has 0 failures
BirdRNG now has 2048 bits of state with an extra 32 of sort-of-state for
a total of 2080, instead of 1056 before. It has no failures in PractRand
testing now.

BeardRNG has some minor bugs fixed as well; nothing changed should
affect its quality, but the changes should improve seeding, especially
re-seeding with setState().