Ramp up quality on BirdRNG; speed is not as good
BirdRNG now has almost no anomalies and no serious anomalies or failures
after testing on multiple seeds. Previously, the anomalies on one seed
turned out to result in errors given some other seeds. Now, quality is
much better, but speed is less-good. After JIT kicks in on int
generation for BirdRNG, it is close to its speed in the previous commit,
but it's not quite there. Long generation now produces a pair of ints
and places them without much change into a 64-bit long, which is slower
than before but should help retain the quality that is the goal here.