Adjust IsaacRNG to match official. Optimize.
The original implementations of IsaacRNG and Isaac32RNG have been moved
to the benchmarks module as IsaacAltRNG and Isaac32AltRNG. The current
IsaacRNG and Isaac32RNG perform slightly better, and IsaacRNG should
match the outputs of the official 64-bit ISAAC generator by Bob Jenkins
(I'm not sure if it does currently, but the old one probably didn't due
to a mistake in translation).

UncommonBenchmark tests the speed of less-tested RNGs, usually ones that
are less tested because they are slower.