BardRNG passes all PractRand tests; little faster
BardRNG is like BirdRNG but has more state, flawless statistical testing
results, and is a tiny bit faster when used directly or much faster when
used indirectly (as the RandomnessSource for an RNG).

BirdRNG changed in this commit to do slightly less unnecessary work per
number; it does change the produced sequence per seed, but so did a fix
to seeding (also in this commit) that now ensures all state elements get
assigned a starting value.