Weird optimizations with Jet; still high-quality
Jet sometimes seems to be improved by the JIT and sometimes seems to be
slowed down after JIT compilation "pessimizes" it. The code is now,
oddly enough, derived approximately from FlapRNG, so the period may be
less-good, but the quality is still excellent, which is surprising
considering Flap has poor quality.