JetRNG now passes PractRand and is much faster
Its period is still uncertain, at the bare minimum pow(2, 32) and
probably more than pow(2, 36) (possibly more by quite a lot), to a max
of pow(2, 512). It is faster than LightRNG, just slightly, at generating
ints, and seems to be the fastest of the int-math generators that use
multiple bits of state while still passing PractRand with no failure
(and also no anomalies for Jet). For GWT usage, it won't be consistent
across browser and desktop/mobile usage due to relying on int overflow,
and BeardRNG or LongPeriodRNG is probably better in the same niche for
GWT, since they use long math.