Add experimental Rule90RNG
I have no idea what the possible period of Rule90RNG is, since it's
based on a non-conventional structure for a PRNG (the Rule 90 cellular
automaton, in 1D, using 64 bits of state as the 1D world). It's not
slow, but not especially fast either; it may be a good technique for a
32-bit RNG. Several documentation mistakes have been corrected in other
classes.