Checkpoint; changing Lunge32 and Thrust32
Both are currently slower than they previously were by a significant
amount, but have excellent quality (no failures or anomalies on 256MB of
data). The version of Thrust32RNG in this commit will be replaced with
the version of Lunge32RNG, and Lunge32RNG will be reverted to its
previous implementation for speed reasons (this previous version's
quality is OK up until you request a large amount of random data, at
which point flaws can be detected). Since Lunge32 and Thrust32 are both
very high-quality now, the faster (which is currently called Lunge32)
will be used to replace Thrust32.