Move Lunge32 to Thrust32, make Lunge32 faster
Lunge32RNG is slightly slower than it was two commits ago, but its
quality is significantly better (no anomalies or failures on 256 MB of
random data). Its period is still problematic at only 2 to the 32,
though. Thrust32RNG now uses the algorithm that Lunge32RNG did in the
last commit, since it is a fair amount faster and how it used its extra
state didn't guarantee much more length for the period.