Update LongPeriod, ThrustAlt for quality
LongPeriodRNG uses the currently-recommended multiplier instead of the
previously-recommended multiplier now.

ThrustAltRNG has gone through intensive editing and is now much closer
in speed to the earlier ThrustRNG, and its quality has kept the
PractRand testing improvement over ThrustRNG (from the last commit) as
well. Its period is a little smaller now, at 2 to the 63 instead of 2 to
the 64, but it should still be useful for the same purposes.

All the RandomnessSource implementations are marked final on the class
now (all of them that I felt could use a speed boost, anyway), which
could have a small improvement on speed. Those classes lacked a reason
to be subclassed directly, unlike RNG, which should be the point to go
from for customized generators.