Improve ThrustAlt constants & determine methods
A slight tweak to the increment used by next(), nextLong(), and skip()
in ThrustAltRNG seems to improve quality a little; I'll see if it's
enough to pass BigCrush without failures once that lengthy test
finishes, but it already does well on PractRand. A different change was
applied to determine() and determineBounded() that allows them to use
all longs as input and have a chance of generating a unique number for
any of them (the period on those has doubled relative to nextLong(), to
2 to the 64, and they each have 2 to the 64 possible inputs). The new
determineBare() method in ThrustAltRNG may help for when speed is
important and the state's increment can be controlled from the calling
code.