Adjust ThrustAltRNG, add IDeterminer interface
ThrustAltRNG was changed so it has a better skip() method, and so it has
a better chance at lasting though large amounts of PractRand testing
(the previous algorithm suffered issues after 1-2TB, while the new one
is at least based on an algorithm that passes 4 TB easily). IDeterminer
may or may not be used in the future, but it provides a way to be able
to swap out determine() methods, which is normally impossible because
they are static.