Major improvements to ThrustAltRNG
The quality seems solid so far through 8TB of PractRand testing. In
100GB of gjrand tests, it had an unbelievable result of 0.981, and on
1TB of gjrand tests (more than the author of gjrand claims to have ever
used), it is almost as good with a result of 0.936, never failing any
gjrand or PractRand tests so far. The determine() method is now slightly
faster than calling nextLong(), which is odd, and may have to do with
the unusual structure of determine(). The determineBare() method didn't
actually yield a performance increase, actually the opposite, so it was
removed since it was a pain to use.