Add a different try at VortexRNG
This version can't use skip() like ThrustAltRNG can, but has a much
larger period (in theory).  Each of the 4 different streams used is
known to pass 100GB of gjrand testing with no failures, and there are 4
reverse-order streams it can also use. If there's a detectable issue
here, it may be fix-able by just XORing the value this returns with the
current stream.