Refactor NLFSR to allow different bit sizes
NLFSR25 is new, and much faster than the NLFSR27, the old algorithm.
NLFSR27 has a somewhat longer period, but not by much. Similar
refactoring may occur in LFSR. This also changes one line in most
RandomnessSource implementations, specializing the return value of
copy() to the actual calling class and not RandomnessSource. This should
help when copy() is actually used on its own, and not as part of RNG's
usage.