Use random choices to break up benchmark patterns
This seems to prevent the deoptimization that hurt Math.sin() and
Math.cos() before, so they perform better than before now. The Math
methods are still not great; even when the approximations do less-well,
there's a 4x speed boost when switching to the approximations. Using
only positive inputs can help NumberTools.cos() and sin() a fair amount
as well, and produces the same outputs if the input is the same modulo 2
* PI.