Speed and GWT compat update for ThrustAlt32
This version still passes PractRand on its full period of 2 to the 32
ints, but is faster than the previous version and avoids anything that
behaves differently on GWT. I believe it's faster because there's more
of a clear path between variables. Before, the state was assigned, a new
variable was made that used the state, and then the returned value
needed both the new variable and the earlier-changed state for its
calculations. Now, the state is assigned, the new variable uses state,
and then the return calculations only use that new variable (state can
be left alone). This is more of a linear path than before, which seems
to optimize well, especially on nextLong(). The GWT emu version has also
been updated in lockstep with the original. The determineSmall() method
has been removed because there's no need for it any more; determine()
will work for all cases that determineSmall() would have made sense
before.