Change VortexRNG to a higher-quality version
This also adds MotorRNG, which is a specialized version of VortexRNG's
algorithm without a changeable stream (it uses one that has been
thoroughly tested and is known to be good). VortexRNG (using the stream
that MotorRNG always uses) passed 32 TB of PractRand testing without any
anomalies and definitely without failures. All of the steps in MotorRNG
are reversible, so with 64 bits of state that cycle through all possible
longs, all possible longs can be returned by its nextLong() method. This
isn't the case for ThrustAltRNG, so there may be some defaults changing
soon.