Changes to Zag32RNG that fix lone PractRand failure
Zag32RNG in the last commit failed the BRank test of PractRand at 4TB.
The usually-well-regarded generator xoroshiro128+ fails that test at
almost all sizes and it's a solid generator, so even though I shouldn't
have been concerned, I decided to try some alternative techniques. With
only minor changes, Zag32RNG has now passed 16TB (and counting) on
PractRand. In 16 or so hours, I will find out if it passes 32TB, which
is the last step. As usual, there's a speed test in the browser at
https://tommyettinger.github.io/SquidLib-Demos/bench/rng/ .