Add Orbit (longer-period ThrustAlt)
OrbitRNG avoids the main flaw ThrustAltRNG has (that it has a fixed set of numbers it can never produce and others it produces more frequently) by extending the period to  a full 2 to the 128 and producing all longs with equal frequency over that period. It is about as fast as Lathe64RNG. Quality tests are fine on it up to at least 8TB; I haven't finished running up to 32TB, but it should be fine because of how well ThrustAltRNG does. Lunge32RNG is in the middle of an experiment; so is SpiralRNG.