Huh, Xara also works as Ara
It seems to sometimes be a fair amount faster when the first XOR is omitted, and the quality seems fine (out of 2TB so far, only one anomaly at 256GB). Xara passes PractRand with only one anomaly at the very end, so if Ara does have issues, I can fall back on Xara.