Starfish aces PractRand
Starfish32RNG passes all 32TB of PractRand with no anomalies! I don't recall any other generators with 2 ints of state doing this well, and this is the only 2D-equidistributed 32-bit generator here that has such a small state size (64 bits). It can be considered maximally-equidistributed as well.

There are some minor experiments in some of the tests that may also be handy.