Overdrive64RNG quality is almost perfect now.
It isn't at all equidistributed, but it passes 32TB of PractRand with no failures or anomalies. I also know the period must be at least 2 to the 47, having brute-forced the CERS sub-generator up to that point without cycling (the CMR sub-generator is still running; it's currently at about 29 times (2 to the 42) generated numbers, and the test ends at 32 times (2 to the 42) numbers). It should be the same speed, since only a rotation amount changed, but we'll see when all the tests are done and I can run benchmarks.