Possible usage of Dirk as LinnormRNG.determine() replacement
I'm still testing this one, but it looks like a decent unary hash may be possible on sequential inputs without mimicking a Weyl sequence (so it won't be multiplying the input by a very large Weyl-style number). This does use two multiplications and two rotates, so it may not be quite optimal on speed, though.