Improvements in MoverCounter64, XoshiroStarPhi32
MoverCounter now appears to be stable as the second-fastest generator in here, and quality so far is excellent. XoshiroStarPhi32 had been tested using two different rotations at the last step, so now it uses the rotation that had 0 anomalies instead of the rotation that had 1 anomaly. The choice of seed could also affect the quality, of course. There's also some minor fixes, such as for MoverCounter64RNG.copy().