Weird speedup on PelicanRNG's normal methods
Some quirk of the translation from near-optimal C to Java caused the code to compile poorly with HotSpot. Time per nextLong() call went from 5.910 ns to 4.590 ns by removing one XOR with a constant (which is mostly relevant for determine() calls, not really for a RandomnessSource), changing a prefix ++ to a postfix ++, and consolidating the XORs with rotations to share an assignment with a multiplication of that XOR result. It's still not especially fast; LightRNG takes 3.977 ns per nextLong() call, and DiverRNG just 3.692 ns. These should all be close enough not to notice unless random number generation is a bottleneck.