PelicanRNG has great quality on determine() and decent speed
PelicanRNG's determine() algorithm is available in DiverRNG as randomize(), since PelicanRNG's algorithm is faster and higher-quality than the variable-shift algorithm used by DiverRNG.determine() before. The new DiverRNG.determine() is identical to LinnormRNG.determine() (with one tweak for determineBounded() ), and randomize() is available when the input sequence might have a pattern that needs to be removed.