Weird results with BumbleBench, more tests added.
BumbleBench gives results that run counter to JMH for some RNGs, maybe because BumbleBench tests a tight loop and JMH tests individual calls. According to BumbleBench, DiverRNG is slower than MiniMoverRNG which is slower than LightRNG, which is about the same as XoRoRNG, and all are slower than ThrustAltRNG. I'm not sure what to make of this.