So it looks like BumbleBench gives very different results
OrbitRNG seems quite fast for a 128-bit generator with a large period, and TangleRNG is like it but faster. What struck me is that Lathe64RNG, the fastest of the edits to xoroshiro128, is still slower than OrbitRNG. Since an even slower variant on xoroshiro is used in GWTRNG (Starfish32), it might be worthwhile to compare Orbit to other variants with a comparable state size. Orbit is being tested again to make sure it passes the TMFn test, which similarly-structured RNGs can be vulnerable to sometimes late in testing.