Benchmark GDX MathUtils vs. SquidLib NumberTools for sin()
Oddly, the fastest of all of them is SquidLib's double-taking-and-returning sin(), by a small degree. MathUtils needed lots of special care to even run in a benchmark because it spends too much time during startup building its LUT, so I had to move the LUT building into a timed area and compensate by doing many more runs of the actually intended-to-benchmark code.