Another quick period test (0xFFFF0000 for 16+16-bit Giblet)
It isn't the longest possible period; that would be 0x100000000. But, Giblet as a 64-bit algorithm has a period of well over 2 to the 127, which would take many lifetimes to exhaust, and unlike xoroshiro128 (regardless of scrambler), it's completely equidistributed (one-dimensionally), producing each 64-bit result 0xFFFFFFFFFFFFFFFF times. Xoroshiro never enters the 0,0 state, so one output is produced less often than all others.