Do some testing on basic MWC generators...
The period on both a 16-bit result and a 8-bit result multiply-with-carry generator is lower than (2 to the (s - 2)), where s is the state size in bits. The highest possible period for a 8-bit-result MWC (with a 16-bit state) is 0x3CFF, using 122 as the multiplier. The output is also distributed very unevenly throughout the 8-bit range.