Fix an issue with Spark8Bit.
It was adding different increments in the first loop and the second loop, which wouldn't happen normally. It has more obvious trouble now; roughly 80% of possible values can't be produced by a given seed, which is much higher than it should be. Sparkle, on the other hand, can't produce 37.5% of values, exactly 3/8, which is right where I think it should be.