Sporty is SplitMix/LaserRandom-like.
It's heavier-duty, with 4 states and the same period, and having effectively 2 to the 192 streams while keeping the same gammas may help. It can produce all outputs over the course of all states, equally often. Over the period of a given stream, it should be unable to produce about 35-40% of possible values, instead of the 80% that Spark can't produce.