Verify what I've seen called "Fermat residue."
This is like a right xorshift, but uses subtraction instead of XOR. It is perfectly distributed for the lower bits, but not evenly-distributed for the upper bits. When computing the Fermat residue on all 16-bit words, every low-order byte appears 256 times, but the high-order byte 0xFF only appears once. All other high-order bytes appear 257 times.