Using the third state as a changing key works.
There's some interesting possibilities here -- the key isn't meant to be recoverable easily, so reversing even a full 3-word output to get the state that produced it might be quite challenging. Going in reverse wouldn't be hard at all, since this just needs to take where it adds to a, b, and c, and change that to subtraction (though it needs to assign q, r, and s first). The Speck steps afterwards would be run normally.