Some interesting findings with an LFSR + counter.
If the counter adds an odd increment AND is incorporated into the 16-bit (reverse direction) LFSR via XOR, then it produces a short period, but if it adds an odd increment *twice* (or just adding 2x an odd number), then it can be incorporated in that way and stay full-period. I don't know if this works for forward-direction LFSRs yet... Or really why it works at all!