Finally found another weird bijection.
Still fiddling with it! This can use the triangular number formula entirely doing math in the normal word-size modular field, and reassemble the MSB based on the status of the "square root bit," here the bit 0x100, or 256 for math modulo 65536. This should use an madd instruction on one register!